pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ui-agent"
version = "0.1.0"
description = "UI Form Testing Agent System based on CrewAI"
requires-python = ">=3.11"
dependencies = [
    "crewai[tools]>=0.86.0",
    "playwright>=1.49.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "openpyxl>=3.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "loguru>=0.7",
    "jinja2>=3.1",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
ui-agent = "src.main:cli"

[tool.setuptools.packages.find]
include = ["src*"]
